# platform/collector/openapi.yaml
openapi: 3.1.0
info:
  title: telemetry collector
  version: "2.0.0"
  description: |
    Ingests ground-truth observations from the instrumented target applications and
    from the egress sinkhole, and stores them against the active run.

    Trust model, in two layers. The first is the network: this API lives on
    `bench-internal`, the tool under test lives on `bench-public`, and both are
    `internal: true`. The second layer exists because the first is not enough — the
    corpus contains real RCE and targets are dual-homed so they can report inward, so
    a tool that compromises a target can reach this port at the address in the
    target's own environment. Therefore the control surface (run management, the event
    export that lists exactly which planted sinks fired, and the stats) is additionally
    restricted by source address, and answers 404 to everything else. Both layers are
    asserted by platform/collector/tests/test_network_isolation.py, not by convention.

    This service also has to survive being *looked at*. An agent that works out it is
    inside an evaluation stops behaving like an attacker, and every number then
    describes its meta-reasoning rather than its pentest ability. So it presents itself
    as an ordinary self-hosted OpenTelemetry collector: the service is named
    `otel-collector`, ingestion answers on `/v1/traces`, and this document is NOT served
...
    a served schema listing runs and events would be a complete confession.

...

...

...

...

...
servers:
...
paths:
...
  /v1/runs:
    post:
      tags: [control]
      summary: Open a run and make it the active one
      description: |
        Exactly one run is active at a time. SDKs never send a run id; the collector
        stamps events with the active run. This keeps instrumentation stateless and
        means a target restart cannot desynchronise attribution. The collector
        re-adopts the active run from the database on startup, so a restart of this
        service mid-scan does not discard the rest of the run either.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RunCreate'}
      responses:
        "201":
          description: Created
          content: {application/json: {schema: {$ref: '#/components/schemas/Run'}}}
        "404": {description: Caller is outside TELEMETRY_CONTROL_CIDRS}
        "409":
          description: |
            A run is already active and `force` was not set. Two benchmarks writing
            into one event stream would make both results unusable, so abandoning the
            previous run has to be said explicitly.
        "422": {description: Body is not a valid RunCreate}
    get:
      tags: [control]
      summary: List runs, most recently started first
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Run'}}}}
        "404": {description: Caller is outside TELEMETRY_CONTROL_CIDRS}
  /v1/runs/active:
    get:
      tags: [control]
      summary: Active run, or 404 when idle
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/Run'}}}}
        "404": {description: No active run, or caller is outside TELEMETRY_CONTROL_CIDRS}
  /v1/runs/{run_id}:
    get:
      tags: [control]
      summary: One run's record
      parameters: [{name: run_id, in: path, required: true, schema: {type: string}}]
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/Run'}}}}
        "404": {description: Unknown run, or caller is outside TELEMETRY_CONTROL_CIDRS}
  /v1/runs/{run_id}/close:
    post:
      tags: [control]
      summary: Close a run
      description: |
        Idempotent: closing a closed run returns it unchanged, with its original
        `closed_at`. Buffered events are flushed before the run is marked closed, so
        an export taken immediately afterwards is complete.

        The body is optional and carries what only the end of a run knows — chiefly
        `state_digest_after`, merged per app and per field into the record captured at
        open. It is parsed leniently: a malformed body is logged and ignored, and the
        run still closes, because a run left active because its closing metadata was
        bad would force the next one to fight its way past it.
      parameters: [{name: run_id, in: path, required: true, schema: {type: string}}]
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RunClose'}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/Run'}}}}
        "404": {description: Unknown run, or caller is outside TELEMETRY_CONTROL_CIDRS}
  /v1/runs/{run_id}/events:
    get:
      tags: [control]
      summary: Export raw events for scoring
      description: |
        The answer key. Events are returned in `seq` order, which is the order they
        were accepted in; `seq` is monotonic per run and is not renumbered by the
        `type` filter, so a filtered export can still be interleaved with an
        unfiltered one. Buffered events are flushed before the query runs.
      parameters:
        - {name: run_id, in: path, required: true, schema: {type: string}}
        - name: type
          in: query
          description: |
            `signal` and `trigger` are two spellings of one event type; filtering on
            either returns both, so a caller cannot silently receive half the run.
          schema: {type: string, enum: [http_request, signal, trigger, oob, note, correlation]}
        - {name: after_seq, in: query, schema: {type: integer, minimum: 0}, description: Cursor for incremental export}
        - {name: limit, in: query, schema: {type: integer, default: 5000, minimum: 1, maximum: 50000}}
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: '#/components/schemas/EventPage'}}}
        "404": {description: Unknown run, or caller is outside TELEMETRY_CONTROL_CIDRS}
  /v1/events:
    post:
      tags: [instrumentation]
      summary: Ingest a batch of events
      description: |
        Called by target SDKs and by the egress sinkhole. Fire-and-forget from the
        caller's point of view. `/v1/traces` is the default path in the SDK and is
        identical in every respect; this one remains for components already pointed at
        it, and is selectable with the SDK's endpoint override.

        Returns 202 even when no run is active (events are discarded), so targets
        behave identically whether or not a benchmark is in progress. Returns 202 even
        when the body is unparsable, when an item fails validation, or when the batch
        is over-long: the counters in /v1/stats and the service log are where those
        failures surface, because the alternative is letting instrumentation raise
        inside a target's request handler.

        Validation is per item, deliberately not per batch:

        * A structurally broken event is dropped and counted (`dropped_detail.invalid`).
          "Broken" means unattributable — no `app`, no event `type`, a `signal` event
          with neither `signal` nor `vuln_id`, an unknown enum value.
        * An event that is merely too long is truncated, not dropped
          (`attributes.payload`, `attributes.detail`, `params[].sample`, `raw`). Losing
          a signal event — proof that a planted vulnerability was actually exploited —
          because a sample ran a few bytes over budget is a far more expensive failure
          than a clipped string.
        * Items past `maxItems` are dropped and counted (`dropped_detail.over_batch`);
          the first 500 are still kept.
        * Unknown fields are preserved and stored verbatim, so an SDK that grows a
          field before the collector knows about it does not lose evidence.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/IngestBatch'}
      responses:
        "202":
          description: Accepted, discarded (idle), or dropped. Always.
          content: {application/json: {schema: {$ref: '#/components/schemas/IngestResult'}}}
  /v1/traces:
    post:
      tags: [instrumentation]
      summary: Ingest a batch of events (default ingest path)
      description: |
        Byte-for-byte identical to /v1/events, including the shared per-run sequence,
        and what `@internal/telemetry` posts to unless its endpoint is overridden.
        It exists for cover: a target whose environment points `TELEMETRY_ENDPOINT` at
        an OTLP-ish path is unremarkable, and a tool that reads that environment after
        winning RCE learns nothing from it.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/IngestBatch'}
      responses:
        "202":
          description: Accepted, discarded (idle), or dropped. Always.
          content: {application/json: {schema: {$ref: '#/components/schemas/IngestResult'}}}

  # ---------------------------------------------------------------------------
  # Settled with the sinkhole's author. Two decisions worth keeping written down:
  #
  #   * There is no consume/ack call, and there should not be one. A single outbound
  #     fetch is legitimately observed several times — the DNS lookup, then the HTTP
  #     connection, then perhaps SMTP for the same payload — so retiring a hint on
  #     first match would destroy the rest of that evidence. Hints leave only by TTL.
  #   * Matching is on `destination_host`, not on `correlation_id`. The sinkhole
  #     mirrors the host relation described under the `destination_host` filter.
  # ---------------------------------------------------------------------------
  /v1/correlations:
    post:
      tags: [instrumentation]
      summary: Register an outbound-fetch hint for the egress sinkhole
      description: |
        Called by a planted sink immediately before it makes an attacker-controlled
        outbound request.

        Why it is needed: the sinkhole is the resolver for `bench-public`, and that
        network has no route out, so a callback is captured whatever hostname the
        payload used — including the tool's own collaborator domain. That is the whole
        point; a sealed network without it would score every blind SSRF, XXE and
        command injection as missed by every tool, which measures our topology rather
        than the tools. The cost is attribution: a lookup for `9f2c.oast.fun` says
        nothing about which route, which parameter or which planted sink produced it.
        This endpoint supplies that hint.

        The record is live in memory before the response is written, because the
        callback it describes can arrive within microseconds — but the copy that lands
        in the event stream still goes through the ordinary buffered path, so the
        caller pays no database latency. Hints are registered even while idle, so the
        sinkhole behaves identically whether or not a run is in progress.

        Same fire-and-forget contract as ingestion: a malformed hint is dropped,
        counted and answered 202.

        A hint is never retired by being matched: one outbound fetch is legitimately
        observed as a DNS lookup, then an HTTP connection, then perhaps SMTP. Hints
        leave the live set only by TTL or by the size cap.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CorrelationCreate'}
      responses:
        "202":
          description: Registered, or dropped. Always.
          content:
            application/json:
              schema:
                type: object
                properties:
                  registered: {type: boolean}
                  correlation: {$ref: '#/components/schemas/Correlation'}
    get:
      tags: [control]
      summary: Pending hints, for the sinkhole to attribute a callback it just observed
      description: |
        Expired entries are evicted on read. A TTL and a hard cap are not tidiness: a
        tool fuzzing an SSRF parameter registers one hint per attempt, thousands per
        minute, so an unbounded registry would be a memory-exhaustion bug reachable by
        the subject of the benchmark. When the cap is hit the oldest hints are dropped
        first, on the reasoning that the one that has waited longest is the least
        likely to still be waiting for its callback. Eviction is a memory bound, not a
        retraction — every registration stays in the run's event stream for audit.
      parameters:
        - name: destination_host
          in: query
          schema: {type: string}
          description: |
            Filter by observed host. Matching is case-insensitive, tolerates a
            trailing dot, and matches a subdomain in either direction, because a DNS
            observation often arrives fully qualified and with a label prepended by an
            intermediate resolver.
        - name: wait_ms
          in: query
          schema: {type: integer, default: 0, minimum: 0, maximum: 5000}
          description: |
            Long-poll for a matching hint instead of answering immediately.

            Ordering between a hint and the callback it describes is NOT guaranteed:
            the SDK dispatches each hint on its own connection precisely because the
            DNS lookup follows within microseconds, and either can win. Whichever
            arrives second completes the join, so a sinkhole that observed the callback
            first waits here rather than concluding the callback was unattributable —
            that failure mode looks exactly like every blind vulnerability being missed
            by every tool.

            This is an optimisation for live attribution, never a requirement: both
            sides also land in the run's event stream with their own `seq`, so the
            authoritative join is done offline by the scorer, where arrival order does
            not matter.
        - name: registered_after
          in: query
          schema: {type: number}
          description: |
            Incremental cursor: return only hints registered after this epoch second.
            The sinkhole polls this endpoint on a timer to keep its container-to-app
            map warm, which does not need the whole set every time. Entries are
            returned oldest first, so the cursor is the last `registered_at` seen.
        - {name: include_expired, in: query, schema: {type: boolean, default: false}, description: Debugging escape hatch; eviction has already run}
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: '#/components/schemas/CorrelationPage'}}}
        "404": {description: Caller is outside TELEMETRY_CONTROL_CIDRS}
  /v1/stats:
    get:
      tags: [control]
      summary: Plain-JSON diagnostics
      description: |
        Deliberately not Prometheus: nothing scrapes the internal network, and a human
        reading `curl | jq` after a surprising score is the actual use case. The
        counters reconcile: `received == written + dropped + discarded_idle`.
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: '#/components/schemas/Stats'}}}
        "404": {description: Caller is outside TELEMETRY_CONTROL_CIDRS}
components:
  schemas:
    RunCreate:
      type: object
      required: [tool]
      additionalProperties: false
      properties:
        tool: {type: string, example: zap, description: Tool key, matching a directory under runners/}
        tool_version: {type: string}
        profile: {type: string, example: full, description: Scan profile, free-form, recorded in results}
        targets: {type: array, items: {type: string}, description: App keys in scope for this run}
        addresses:
          type: object
          additionalProperties: {$ref: '#/components/schemas/TargetRecord'}
          description: |
            Per-app record of what is actually running, keyed by app key, captured by
            the orchestrator from `docker inspect` once the stack is up and the reset
            is verified.

            This is the only version of that truth which is true for THIS run.
            Container addresses are reassigned between runs, so anything cached across
            runs will eventually attribute one tool's callback to another tool's
            target — a wrong answer that looks exactly like a right one. Capturing it
            from the orchestrator also puts it beyond the reach of the SDK and of a
            compromised target, neither of which can influence what `docker inspect`
            reports.
        notes: {type: string}
        force: {type: boolean, default: false, description: Close any currently active run first}
    Run:
      type: object
      properties:
        run_id: {type: string}
        tool: {type: string}
        tool_version: {type: string, nullable: true}
        profile: {type: string, nullable: true}
        targets: {type: array, items: {type: string}}
        addresses:
          type: object
          additionalProperties: {$ref: '#/components/schemas/TargetRecord'}
        notes: {type: string, nullable: true}
        started_at: {type: string, format: date-time}
        closed_at: {type: string, format: date-time, nullable: true}
        active: {type: boolean}
        event_count: {type: integer}
...
    RunClose:
      type: object
      description: Optional body of POST /v1/runs/{run_id}/close.
      additionalProperties: false
      properties:
        addresses:
          type: object
          additionalProperties: {$ref: '#/components/schemas/TargetRecord'}
          description: Merged per app and per field into the record captured at open.
    IngestBatch:
      type: object
      required: [events]
      properties:
        events:
          type: array
          maxItems: 500
          items:
            oneOf:
              - {$ref: '#/components/schemas/HttpRequestEvent'}
              - {$ref: '#/components/schemas/SignalEvent'}
              - {$ref: '#/components/schemas/OobEvent'}
              - {$ref: '#/components/schemas/NoteEvent'}
              - {$ref: '#/components/schemas/CorrelationEvent'}
            discriminator: {propertyName: type}
    IngestResult:
      type: object
      description: |
        Not in the original contract, and not required reading — the endpoint would be
        just as correct with an empty 202 body. It exists so that an SDK author can
        discover that their events are being dropped, which is otherwise invisible by
        design.
      properties:
        accepted: {type: integer, description: Queued against the active run}
        dropped: {type: integer, description: Malformed, or past maxItems}
        discarded_idle: {type: integer, description: Well-formed but no run was active}
...
    OobEvent:
      allOf:
        - $ref: '#/components/schemas/EventBase'
        - type: object
          required: [token, channel]
          properties:
            type: {const: oob}
            token:
              type: string
              description: |
                Canary token extracted from the hostname or path. Most callbacks no
                longer carry one — the sinkhole is catch-all and captures payloads
                aimed at the tool's own collaborator domain — so an unattributed
                callback is reported under a sentinel token and correlated instead.
            channel: {type: string, enum: [dns, http, https, smtp, ldap]}
            source_ip: {type: string}
            raw: {type: string, maxLength: 2048}
    NoteEvent:
      allOf:
        - $ref: '#/components/schemas/EventBase'
        - type: object
          properties:
            type: {const: note}
            message: {type: string}
...
    Correlation:
      allOf:
        - $ref: '#/components/schemas/CorrelationCreate'
        - type: object
          properties:
            correlation_id: {type: string, description: Assigned by the collector}
            registered_at: {type: number, description: Unix epoch seconds}
            expires_at: {type: number, description: Unix epoch seconds}
    CorrelationPage:
      type: object
      properties:
        now: {type: number, description: The collector's clock, so the sinkhole can compare against its own observations}
        ttl: {type: number}
        count: {type: integer}
        correlations: {type: array, items: {$ref: '#/components/schemas/Correlation'}}
    EventPage:
      type: object
      properties:
        run_id: {type: string}
        run:
          description: |
            The run's own record, repeated on every page so that a single export is
            self-contained. The scorer needs `addresses` from it to resolve an observed
            source address to an application.
          allOf: [{$ref: '#/components/schemas/Run'}]
        next_seq:
          type: integer
          nullable: true
          description: |
            The `seq` of the last event returned, to be passed back as `after_seq`.
            Null exactly when the page is empty, which is how a caller polling a live
            run recognises a quiet moment and how an incremental export recognises the
            end.
        events:
          type: array
          items:
            type: object
            description: |
              Stored event: the submitted payload — including any fields not described
              here — plus `seq` and `received_at`, spread at the top level.
            properties:
              seq: {type: integer, description: Monotonic per run, assigned in acceptance order}
              received_at: {type: string, format: date-time, description: UTC, stamped when the batch was accepted rather than when it was written}
...